Idling: check existing loadbalancer before moving the vip to the idling
Right now we add the vipProtocol to the list of vipProtocols to remove
only if the loadbalancer is not the idling one. The problem is, we
update the loadbalancer just before this check, with the effect of not
removing the vip from the existing balancers. This causes the vip to
stay both on the idling and non idling loadbalancers, with non
predictable effects.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>