Fix duplicate incrementing of subnet allocation metric
When the master is restarted, all the nodes are fed using
both a fetch of existing nodes and by the add node event.
In case the node already had subnet annotations, we were
updating the subnets allocated metric again. This commit
checks if new subnets were allocated and only then calls
the code to update metrics associated with allocated
subnets.

Signed-off-by: Aniket Bhat <anbhat@redhat.com>