Simplify and use hairpin_snat_ip for Cluster Lbs
Currently when traffic is sent to an OVN Lb and is hairpinned
back to the same dstIP as the srcIP the OVN lb randomly chooses
and IP for the SNAT.  This makes it hard for the network policy
code to accurately allow traffic in some scenarios.

This Commit forces the IP chosed for SNAT to be 169.254.169.3 for
ipv4 and fd69::3 for ipv6.  Which will allow us to accurately
allow traffic with network policy without creating a dependency
between the network policy and service objects within OVN-K8s.

The new option is only applied to the ovn cluster Loadbalancers
because we only can apply network policy to cluster networked pods.

It also moves to creating and getting all OVN-K loadbalancers with
the loabalancer package to eliminate running spurious nbctl lb commands
all across the codebase. Previously we were just making the k8's
cluster loadbalancer in the `loadbalancer` package, now we make OVN-K
lbs there:

1. K8's cluster LBs
2. GR LBs
2. K8's worker LBs

Lastly it ensures that the correct options are applied on update for
the k8's cluster LB

Signed-off-by: astoycos <astoycos@redhat.com>