Infer subnet for node /128 IPv6 addresses
When node addresses are allocated through DHCPv6 IA_NA, they might get
a /128 prefix address. These node addresses are set on the gateway
router port as well. OVN routers infers on-link routes from subnet
information on the address itself. Since /128 addresses are not on a
subnet, cluster traffic is routed through the node gatway instead of
directly.

Note that OVN does not support learning routes from ICMPv6 route
advertisement and does not support adding on-link static routes.

This patch infers the subnet of the interface looking at the node
routing information and replaces the /128 prefix with the subnet
prefix.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>