Fix node handler stepping on pod handler
The code from AddNode to ensure pods already on that node were added
back was flawed. The two goroutines could be trying to add the pods at
the same time, which could lead to errors. This fixes it by utilizing
the retryPod goroutine, which is thread safe with the pod handler.

Therfore now the AddNode will simply add the pods to the retry map, and
request retry to run immediately. This isolates the pod add logic back
to the pod handler itself.

Fixes: #2281

Signed-off-by: Tim Rozet <trozet@redhat.com>