acl-logging: react to namespace annotations
With this commit, we now activate acl logging(with a certain
configurable threshold) when the namespaces are annotated, or stop
logging when the annotation is removed.

The code to configure ACL logging for the default-deny ACLs per
each individual namespace was duplicate; this commit changes this
update method to work based on parametrized port group UUID and
suffix to achieve the same.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>