acl-logging, e2e: assert the ACL log severity on net policies
Two new e2e tests are added; both of which use a single deny all
policy. In the first test we check the initial logging severity
is the configured one, and in the second test, we update the
logging severity from the original one, and assure the updated
logging severity is the one found in the container logs.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>