Fix GetPortAddresses for HBO
In GetPortAddresses, we check if we have both PortMAC and IP for
the port. If not we return nil for both values. While this works
for normal ports, for the HBO ports which don't have IPs for
the int-<node> interface, we will always hit this condition and
return an empty PortMAC which in turn triggers the lsp-add even
if the portMAC already exists.

When we have an external entity updating the node object, the
watcher triggers handleOverlayPort and we keep hitting this
condition.

Let's remove the "len(addresses)<2" check since its not needed
for HBO and in normal pod port case, pods.getPortAddresses already
handles the case for empty podMac or podIPs. This way when portMAC
is set but not the IPs, we will still return the portMAC and the
respective code paths handle it as needed.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>