Refactor - Change Smart-NIC to DPU
The correct term for such devices is a
data processing unit (DPU) as the hardware
contains general purpose CPU in addition to
networking hardware and offload engines.

The term Smart-NIC is usually referred to NICs
that have some programmable networking logic that
allows it to offload networking rules such as
Openflow rules or Linux TC rules.

A DPU is a Smart-NIC but a Smart-NIC
is not necessarily a DPU.

An example of such NICs that fall in the category of
Smart-NICs but are not DPU:
- Nvidia Mellanox ConnectX network adapters
- Netronome Agilio network adapters
- Intel FPGA based adapters

Note:
As there are no known deployments in production
of OVN-Kubernetes with DPUs both from upstream and
downstream forks e.g for Nvidia/RedHat products, as
well as official release of OVN-Kubernetes with this
change. This commit (for the sake of code simplicitly)
does not include backward compatibiliy for handling
the modified DPU specific pod annotation and
configuration flags.

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>