Service: move to per-service loadbalancers, implement affinity
This removes all of the complexity around managing vip membership.
Instead, it creates one or more load balancers per service.
Additionally, it tags those load balancers with the GroupKind and
NamespacedName of the "owning" Kubernetes object, so we can easily
find stale objects and reconcile.

It includes an idempotent load-balance reconciliation engine, so we can
generate just the ideal state in the ServiceController and let the
reconciler take care of the details. This includes a read-through cache,
to minimize unnecessary reads from the database.

A much-simpler repair engine handles cleaning up any stale Load_Balancer
objects on startup.

This also implements session affinity, which is just an option on the
Load_Balancer object.

Signed-off-by: Casey Callendrello <cdc@redhat.com>