Metrics: move e2e timestamp to nb_cfg column, scrape chassis nb_cfg
By moving the e2e timestamp metric to nb_cfg, we get an additional
assurance of end-to-end connectivity: the ovn-controller writes this
value back to the sbdb.

This means we can scrape the nb_cfg from the Chassis_Private table and
determine if there are any ovn-controllers that are runnin behind.

Signed-off-by: Casey Callendrello <cdc@redhat.com>