Fix lb delete during node deletion
When a node gets deleted, the logical switch and
gateway router get removed first. However we do
not update the LBCache and this causes dependency
failures during deletion leaving behind wrong ip->vip
pairing in the lbs.

When the pods on this node go away, ensureLBs
gets triggered and it tries to update endpoints across
load balancers. This set load_balancer command is
batched with ls-lb-del & ls-lr-del commands to remove
association of the lbs from the removed switches and
routers. These commands fail together because the switches
and routers are not present in ovn.

This PR updates the LBcache as soon as we remove the
routers and switches.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>