Re-enable e2e test without disable-snat-multiple..
PR #2376 removed a bunch of CI runs, overlooking the fact that it
happened to remove the CI run where `disable-snat-multiple-gws`
is enabled. This jobs **must** continue to run because it tests the
egress IP functionality. Egress IP and disable-snat.. are mutually
exclusive, but since both options are supported by ovnkube, both need to
be validated in CI!

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>