Reconnect to db if there is transaction error
Under heavy load, ovsdb monitor gets dropped silently and we
stop receiving update messages from ovsdbserver. This leads
to stale cache and inconsistency with what's in the DB
leading to transaction errors. While this is originally
a ovsdbserver bug, a workwround fix here would be to
trigger reconnect to recreate the monitor and start
receiving update messages, when a transaction error occurs.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>
Co-Authored-By: trozet@redhat.com