return on error while parsing node's subnet annotation
on node addition to a cluster, say we failed in performing gateway setup.
the node is added to gatewayFailed store. on next node update,
syncNodeGateway() is called and it could be that the node still doesn't have
host-subnets node annotation set and we will return nil from that function
and remove the node from the gatewayFailed store without setting up gateway.

Reported-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>