Change the IP addresses used for checking if an egress node is reachable
To check if a particular node is reachable ovn-kubernetes is using the
node's primary IP to open a TCP connection on port 9.
This is not going to work if the cluster doesn't allow for this type of
traffic.

To avoid connecting to the node's primary IP use one of the node's
logical switch management port addresses.

Signed-off-by: Patryk Diak <pdiak@redhat.com>