Create bfd sessions and feed it into lr-route-add
We have a bug in OVN: https://bugzilla.redhat.com/show_bug.cgi?id=2007549
due to which bfd sessions are not established correctly. Due to this
if we have x app pods across y nodes served by an exgw pod, we'd expect
y bfd sessions against that exgw pod but currently only 1 session from
the node where the first app pod was spawned will be created.

Until the OVN bug is fixed and consumed, we'd need a workaround in ovn-k
whereby we create the bfd session ourselves if it doesn't exist and feed
it into the lr-route-add command.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>