Node:set isOvnUpEnabled in full and smart-nic mode
Today when ovnkube node is in smart-nic mode
it does not check if OVN supports setting of
OVS.Interface.external-id:ovn-installed and
OVNSB.Port_Binding.up via getOVNIfUpCheckMode()
which is used to check if an ovs port was
successfully programmed.

It instead assumes this is not supported and
falls back to legacy flow check (cni.doPodFlowsExist).

This commit calls getOVNIfUpCheckMode() in both
ovnkube node modes. For the third mode, smart-nic-host,
there is no need to call this method as ovs is not used
on the host.

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>