CARRY: pods: remove unnecessary LSPGet() calls
LSPGet() already returns the LSP's Options so just copy them
instead of calling LSPGetOptions() which has to lock the go-ovn
cache and look up the LSP.

Second, change the after-OVN-execute LSPGet() to LSPGetUUID()
after reading the UUID from the OVN execute response. go-ovn
uses the UUID as a cache index, so LSPGetUUID() is much faster
than getting an LSP by name which requires walking the cache
row-by-row.

Third, eliminate the LSPGet() that GetPortAddresses() does.
We already know the addresses from the first LSPGet() so we
only need to parse them. If the port didn't exist in the
first place, we know there are no addresses to parse so
don't even bother.