iface-id-ver feature when disabled must be disabled on both master and node
currently, the iface-id-ver feature is disabled on the node alone. so on
the master side we create LSP with iface-id-ver set and on the node side
we are not setting external_ids:iface-id-ver on the OVS port. this results
in ovn-controller not claiming the logical port and performing the physical
binding.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>