Support configurable ovn-monitor-all.
With commit 09af653743 "Use DGP to connect logical switches to the
cluster router.", each node only cares about a very small portion of the
SB DB. However, currently ovn-monitor-all is always set to true, which
means each node still monitors the whole SB DB data.  To fully benefit
from the scale advantage of the commit 09af653743, it is better
to set ovn-monitor-all to false, so that each node conditionally
monitors the data relevant to the node only, which would largely reduce
the data needed to be transfered from SB DB to each node, which in turn
reduces the time spent on ovn-controller initial computing, and also
largely reduce the memory footprint on each node.

Although the benefit of conditional monitoring on the ovn-controller
side is obvious, the impact on SB DB server side is mixed. It reduces
the cost on the jsonrpc streams because the amount of data to transfer
to each client reduces significantly, but the condition filtering cost
increases. So it is unknown for now whether it is a performance gain or
a loss for the SB server side when setting ovn-monitor-all to false on
ovn-controller side.

Due to the above considerations, this patch makes the setting
configurable, and keep the default behavior unchanged for now. In the
future if conditional monitoring is desired for majority of use cases we
can change it to the default.

Signed-off-by: Han Zhou <hzhou@ovn.org>