ovn-db-checker: don't use ovsdb-tool for sid
Using ovsdb-tool directly on the database file may cause high memory
usage. This happens because ovsdb-tool has to read the first record in
the database in order to get the servier id (sid). In some cases,
database compaction may have occurred and the first record in the DB may
be giant. This causes periodic spikes of ovn-db-checker to over 1GB of
RAM.

This patch migrates ovsdb-tool to ovn-appctl for getting the sid during
periodic checks. For initial bring up it's fine to keep using ovsdb-tool
as it only happens at pod start.

Signed-off-by: Tim Rozet <trozet@redhat.com>