acl-logging, tests: fix default deny all race
The deny all ACLs are associated to a namespace; despite that, their
name features the policy that triggered their creation.

The default deny all ACLs (and port groups) are created for the first
policy being provisioned in the namespace; the tests were assuming that
the first policy being iterated would be the first one created, something
which proved to not be deterministic.

By provisioning an initial policy, we *know* deterministically which
policy created the default deny ACLs and port groups, and as such,
can match against them.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>