ovn-dbmanager: make db errors/retries more obvious
The logic for handling db retries was embedded in the checks themselves,
and the retry counter variables were shared across the functions. This
is confusing as the global variables are used in different ways in the
functions to determine if there is a database failure.

This changes the checks to instead of modifying the retry counter/raft
reset, they simply return an error. If the error is of type DBError then
the main function which runs all checks will determine if the counter/DB
needs to be reset.

Also fixes a minor bug in the ensureElectionTimeout function where DB
was never getting reset correctly.

Signed-off-by: Tim Rozet <trozet@redhat.com>