CARRY: go-ovn: prevent deadlock processing Updates during initial DB dump
It appears that ovsdb-server will sometimes send updates after we register
our database monitors, but before we've received and processed the initial
database dump. This causes a deadlock because during the initial monitor
setup we must lock the database cache, but the update processing also
locks the database cache.

Avoid this by making the notifier defer updates during the initial
database dump, and processing them only after the dump is done.