fix node connectivity to service backed by egress IP pods
The host network couldn't connect to services backed by egress IP
matching pods since the return packets would get forwarded to the egress
nodes for these pods because of the routing policies specified for the
egress matching pods. This would in turn break the connection. Hence
what we need is to specify routing policies which forces the packet to
go to the originating GR instead of getting routed to the egress node
GR.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>