Enable connectivity to the host network for egress firewall matching pods
Egress firewall matching pods didn't currently have any connectivity to
the host network when a "deny 0.0.0.0" rule is created. That type of a
rule blocked access to all cluster nodes. This was not in-line with the
expectation of how the feature was working for people coming from
openshift-sdn. They thus needed to explicitly allow all node IPs if they
required pods to access the Kube API service, for example. That is not
optimal though as if nodes leave / join the cluster, the rules need to
be updated by the admin. We now have documentation specifically
mentioning that we don't support this, so I am not sure we want this:
https://github.com/openshift/openshift-docs/pull/35311 but this patch
was pretty straightforward to code, so we can discuss it on the PR.
Note: anyone using ovn-kubernetes before this patch (and setting up
specific allow rules for the nodes) should not be impacted negatively by
this patch...but they will have a duplicate ACL for every node they've
explicitly allowed

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>