Scale: Don't trigger updateNode unless needed
Currently any update to any of the node fields
triggers an `updateNode` from the updateFunc of
node watcher. For each node, every 5mins there
is `updateNode` being triggered since something
seems to be updating the node object. In reality
`updateNode` only needs to be called if node.Name
or switchName or gwRouterName or hostSubnet
annotation or L3Gateway annotation change (which
are rare changes) and it has no effect on other
unrelated field changes.

In large clusters, even if its event-triggered it
will cause a lot of noise because mostly a lot
of node fields might be getting updated that have
no relation to what `updateNode` is actually doing.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>