NetworkManager running on SmartNIC creates more than one system ports
so, with NM managing the SmartNIC both `p0` and `pf0hpf` is marked as
system port and the GetNICName() returns `pf0hpf` as the uplink port.
this results in incorrect OpenFlow flows being configured on the OVS
bridge. so, skip this method if there are more than one system Type
ports and fallback to other methods.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>