Fix pod handlers for egress IP
The EgressIP object has a namespaceSelector, which means
we can match on multiple namespaces. Each namespace will
need a individual watcher for the pods in that namespace.
The podHandlerCache is what takes care of tracking those
watchers. Now before: that cache stored it's watchers on
a per-egress-ip basis, which is wrong because if we have
multiple namespaces matching the egress IP we stop
tracking all namespaces except the last one which is
added to that cache. This means that if an egress IP is
deleted we don't clean up any namespace watcher except
the last one added and we will still continue reacting to
events for them and setting up egress IPs

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>