egressgw: only return unique elements from getRouteInfosForGateway()
If a single routeInfo had multiple external routes that matched the
same gateway IP, that routeInfo could be added to the returned list
multiple times. It was only locked once, however.

But deletePodGWRoutesForNamespace() iterates over the list returned
by getRouteInfosForGateway() and calls Unlock() on each element in
the list, which can create an unbalanced locking situation (multiple
Unlocks for a single Lock) in the above situation where a routeInfo
has multiple external routes matching the same gatewayIP.

Signed-off-by: Dan Williams <dcbw@redhat.com>