Support upgrade to newLGW
We add a new topology version called
routingViaHost, where we take a major
step in bridging the gap between LGW and
SGW modes. This new topology will not have
br-local, DGP, gw0. It will have all the
new openflows on breth0 as for SGW to
handle ingress/service traffic.

At this point, the only differences between
LGW and SGW are:

1) egress traffic from pod will still leave
via mp0 in LGW while it leaves via GR in SGW.
2) acls for egress-firewall are on node switches
in LGW while they are on the join switches in SGW.
3) 501 hybrid policy is used to steer traffic
towards GR in LGW while its not needed in SGW.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>