isEquivalentNAT: Allow EIP to be empty for SNAT
In the case where we are trying to find if nat.ExternalIP
is the same for existing and searched nats, we need to
compare the field only if its not empty.

This patch fixes the match condition for externalIPs so that
they behave as expected. Else they break dnat_and_snat/dnat
deletions.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>