Use OnModelUpdates for changing ovn-topo-ver
Using OnModelMutate doesn't make changes to existing
fields in the map. It simply inserts or deletes stuff.
We need to use OnModelUpdate to change the externalIDs
field in ovn_cluster_router instead.

To ensure OnModelUpdate doesn't wipe out existing fields
in the map, we need to pull the latest state from
libovsb cache and then update the k8s-ovn-topo-version.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>