Create ETP=local LBs
In SGW, we create special skip_snat=true LBs
for NP,EIP&LB services. Let's do the same for
LGW to maintain parity even if ETP=local traffic
would go in via mp0.

We are also replacing the vip on NodePort svc per
node switch LB to the special masueradeIP to
support ETP=local on LGW. It's done for both
modes although we won't use this in SGW.

This takes us closer in bridging the gap between
the two modes and making it more like
routingViaHost or not instead of two different "modes".

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>