Add retry for libovsdb transact when not connected
With recent changes to libovsdb, transactions will be blocked during a
state of disconnect. Libovsdb handles this for us. However, there are
periods of time where a network disruption, or server side disconnection
can occur. During this time libovsdb may not have recognized the
connection is down yet, and may attempt a transaction that will result
in "not connected" error. This compensates for it, by adding retry logic
in ovn-kube as well.

Signed-off-by: Tim Rozet <trozet@redhat.com>