gatewayinit: Reduce dSNAT difference btw S&LGWs
dSNAT has been false on LGW since egress traffic
doesn't use the GR path. In SGW, this could be true
in which case we'd add individual nat's for pods
not belonging to the routing-external-gws pod.

Let's remove the check for gateway modes and
add the clusterSubnet SNAT irrespective of the modes
when dSNAT=false.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>