egressgw: fix deadlock due to behavior of RWLock
1. ensureRouteInfoLocked() locks the cache
2. getRouteInfosForGateway() tries to grab the cache RLock() but is blocked on 1
3. ensureRouteInfoLocked() gets the routeInfo from the cache releases the cache lock
4. getRouteInfosForGateway() acquires the RLock() on the cache and begins iterating routeInfos
5. ensureRouteInfoLocked() locks the routeInfo
6. another call to ensureRouteInfoLocked() tries to acquire the cache lock, this blocks all new acquisitions of the cache RLock() per the RWLock documentation
7. the first ensureRouteInfoLocked() now tries to acquire the cache RLock() to validate that its routeInfo is still in the cache, but is blocked on 6
8. getRouteInfosForGateway() tries to lock the same routeInfo that 5 has already locked; never releases the cache lock

The core problem is holding the routeInfo lock while trying to acquire the
route cache lock, which leads to this deadlock. The only reason we try to
do that is to ensure we don't use a deleted routeInfo object, but we can
track deletion inside the routeInfo itself instead.

This solution fixes the deadlock but does leave a small window of time
where, if a gateway pod is created/deleted/created/etc quickly the
operation will return an error (because the pod's routeInfo is marked
deleted but not yet removed from the cache) but that's better than
a deadlock since we can work around it by leaving a bit of time between
the delete/re-create operations.

Signed-off-by: Dan Williams <dcbw@redhat.com>