NATs not getting deleted on GWR
When disableSNATMultipleGWs=true, per pod
SNATs are added for normal pods and for the
served pods, these snats are deleted. This
SNAT deletion wasn't working because we
were passing the podIP/hostsubnetcidrmask
instead of the expected podIP/32 which is
done during the snat addition. Hence the
right nats were not getting matched and
thus stale nat entries were left behind
in the GWR + external gateway tests were
failing since they expected the traffic
from the served pods to not be snated.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>