Make egressIPs compatible with externalgws
Currently we use the disableSNATMultipleGWs
option for externalgws feature to ensure we
do not SNAT podIPs to the nodeIP. This means
we enable a per pod snat for the pods not
managed by externalgws.

EgressIPs feature can work only if the SNAT
for an egressIP managed pod is to the egressIP
and not the nodeIP as is the case when
disableSNATMultipleGWs is set.

So far we were declaring these features as
mutually exclusive. Let's stop that and instead
add the logic to simply delete SNAT to nodeIP
for an egressIP managed pod & add SNAT to nodeIP
for an egressIP un-managed pod.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>